Add PASTE_IGNORE_COMMENTS.
2001-05-08  Jody Goldberg <jgoldberg@home.com>

	* src/clipboard.c (clipboard_paste_region) : Add PASTE_IGNORE_COMMENTS.

	* src/commands.c (cmd_merge_cells_undo) : Use it here.

2001-05-07  Jody Goldberg <jgoldberg@home.com>

	* src/sheet-merge.c (sheet_merge_remove) : adjust comments.
	(sheet_merge_add) : ditto.

	* src/sheet-object-cell-comment.c (comment_get_points) : handle merged
	  cells.

	* src/commands.c (cmd_unmerge_cells) : don't queue an undo object if
	  there is nothing to do.
	(cmd_merge_cells) : filter existing ranges too.

+++ plugins/sc/sc.c
@@ -315,7 +315,7 @@ sc_parse_let (sc_file_state_t *src, const char *cmd, const char *str, int col, i
 
 typedef struct {
 	const char *name;
-	size_t namelen;
+	int namelen;
 	gboolean (*handler) (sc_file_state_t *src, const char *name,
 			     const char *str, int col, int row);
 	unsigned have_coord : 1;
